+++ Programs/Sources/Libraries/mcon/Vector/main.cpp
@@ -13,10 +13,10 @@ void DumpMatrix(mcon::Matrix<T>&m, const char* fmt = NULL)
     else if (typeid(double) == typeid(T))
     {
     }
-    for (uint i = 0; i < m.GetRowLength(); ++i)
+    for (size_t i = 0; i < m.GetRowLength(); ++i)
     {
         printf("| ");
-        for (uint j = 0; j < m.GetColumnLength(); ++j)
+        for (size_t j = 0; j < m.GetColumnLength(); ++j)
         {
             printf("%g", static_cast<TestType>(m[i][j]));
             printf("\t");
@@ -30,8 +30,8 @@ extern void test_matrix_api();
 
 static void test_matrix_determinant(void)
 {
-    const uint numArray = 4;
-    const uint numData= 4;
+    const size_t numArray = 4;
+    const size_t numData= 4;
     mcon::Matrix<TestType> mat1(numArray, numData);
     mat1[0][0] = 1;
     mat1[0][1] = 2;
@@ -56,8 +56,8 @@ static void test_matrix_determinant(void)
 
 static void test_matrix_inverse(void)
 {
-    const uint numArray = 4;
-    const uint numData= 4;
+    const size_t numArray = 4;
+    const size_t numData= 4;
     mcon::Matrix<TestType> mat1(numArray, numData);
 #if 0
     mat1[0][0] = 1;
@@ -108,12 +108,12 @@ static void test_matrix_inverse(void)
 
 static void test_matrix_multiply(void)
 {
-    const uint numArray = 3;
-    const uint numData= 4;
+    const size_t numArray = 3;
+    const size_t numData= 4;
     mcon::Matrix<TestType> mat1(numArray, numData);
-    for (uint c = 1, i = 0; i < numArray; ++i)
+    for (size_t c = 1, i = 0; i < numArray; ++i)
     {
-        for (uint j = 0; j < numData; ++j, ++c)
+        for (size_t j = 0; j < numData; ++j, ++c)
         {
             mat1[i][j] = c;
         }
@@ -135,13 +135,13 @@ static void test_matrix_multiply(void)
 
 static void test_transpose(void)
 {
-    const uint numArray = 3;
-    const uint numData= 5;
+    const size_t numArray = 3;
+    const size_t numData= 5;
     mcon::Matrix<TestType> mat(numArray, numData);
-    for (uint i = 0; i < numArray; ++i)
+    for (size_t i = 0; i < numArray; ++i)
     {
-        printf("mat [%d, 0 .. %d] ", i, numData-1);
-        for (uint j = 0; j < numData; ++j)
+        printf("mat [%d, 0 .. %d] ", static_cast<int32_t>(i), static_cast<int32_t>(numData-1));
+        for (size_t j = 0; j < numData; ++j)
         {
             mat[i][j] = (i+1)*10+(j+1);
             printf("%f ", mat[i][j]);
@@ -151,10 +151,10 @@ static void test_transpose(void)
 
     mcon::Matrix<TestType> matt(1,1);
     matt = mat.Transpose();
-    for (uint i = 0; i < matt.GetRowLength(); ++i)
+    for (size_t i = 0; i < matt.GetRowLength(); ++i)
     {
-        printf("mat [%d,0-%d] ", i, matt.GetColumnLength()-1);
-        for (uint j = 0; j < matt.GetColumnLength(); ++j)
+        printf("mat [%d,0-%d] ", static_cast<int32_t>(i), static_cast<int32_t>(matt.GetColumnLength()-1));
+        for (size_t j = 0; j < matt.GetColumnLength(); ++j)
         {
             printf("%f ", matt[i][j]);
         }
